+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Improved code formatting for expressions within `WHERE` clause [#1421](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1421)
 - Added the parameters panel for `FlexibleSearch` editor [#1425](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1425)
 - Improved scroll for parameters panel and parameters representation [#1426](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1426)
+- Added Date Picker for Date type parameters [#1429](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1429)
 
 ### `Groovy` enhancements
 - Adjusted execution mode selection `commit`/`rollback` [#1422](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1422)